move tests related to reading/writing files to own file
